+++ packages/unrag/registry/connectors/notion/index.ts
@@ -1,0 +1,7 @@
+export { createNotionClient } from "./client";
+export { normalizeNotionId32, normalizeNotionPageId32, toUuidHyphenated } from "./ids";
+export { renderNotionBlocksToText } from "./render";
+export { buildNotionPageIngestInput, loadNotionPageDocument, syncNotionPages } from "./sync";
+export * from "./types";
+
+